+++ lib/augmented.rb
@@ -2,6 +2,7 @@ require 'augmented/version'
 
 require 'augmented/arrays'
 require 'augmented/enumerators'
+require 'augmented/exceptions'
 require 'augmented/hashes'
 require 'augmented/modules'
 require 'augmented/objects'
@@ -12,6 +13,7 @@ require 'augmented/symbols'
 module Augmented
   include Arrays
   include Enumerators
+  include Exceptions
   include Hashes
   include Modules
   include Objects