+++ lib/augmented/exceptions.rb
@@ -1,0 +1,7 @@
+require 'augmented/exceptions/detailed'
+
+module Augmented
+  module Exceptions
+    include Detailed
+  end
+end