+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Added `Exception#details`, `Exception#details=`, `Exception#detailed`
+
 ## [0.2.3] - 2021-05-29
 
 - Added `String#truncate`, `String#truncate!` and `String#blank?`